requirements.txt
numpy==1.21.2
requests==2.27.1
Flask==3.0.0
gunicorn==20.1.0
python-dotenv==1.0.0
pandas==1.2.0
